requirements.txt
# Progress Bar
tqdm

#.Env Secrets
python-dotenv

#Scraping
requests
lxml
beautifulsoup4

#Parsing
pydantic
google-genai
